another redundant promise
the buildChart function isn't doing anything asynchronous, so it can
can just return the config (instead of returning a promise for the
config).

NB. this change makes no difference to the calling code in the other
file. that code functions exactly the same before and after this
commit. The 'then' call in the other file will always keep unwrapping
promises until it gets something that is not a promise.