Chart sizes redux, add option to remove text
At smaller screen sizes (iPad portrait and smaller) chart dimensions
are optimized for common fixed sizes (e.g. iPhone 6, Nexus etc) rather
than arbitrary browser widths. Each size is fluid up to a point, so a
chart won’t scale to more than about 1.09 times it’s actual size,
although is can scale down  a bit more. Similarly heights are roughly
tweaked to the known device sizes - while trying to maintain a decent
aspect ratio for this kind of data - so in landscape orientation the
chart never spans more than the screen. The height values may still
need to be tweaked a little bit.

The chart drawing option removes the heading, subhead and source from
the SVG (so it can be added to the HTML). Axis ticks and plot
annotations are still rendered as part of the SVG.