Analytics: add o-tracking, set up new GA property
Sends Analytics to Spoor and Lantern.

Includes a “microsite_name” field of “us-elections” which allows for
easier custom queries on the Spoor platform.

This also changes the Google Analytics UA code to use a Property set up
specifically for this microsite and includes link tracking and
demographic data.

The client side tracking JS is temporary. It works well enough but I’m
going to refactor client JS loading so that it’s asynchronous, once
I’ve done that I’ll tidy this code up.