Monitor the status of data processing
This adds an endpoint that allows you to inspect problems with the code
the processes the spreadsheet.

It also allows you to see these issues in the application debug logs.

The reason we need this is that the result service will throw errors if
it notices problems with the data coming into the service. This is a
safety mechanism to prevent bad data showing up on the website and give
people time to fix it.

When an error is thrown the service continues to use the data from the
cache, ie. the last time it successfully processed the spreadsheet.